query_manager/requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bcrypt>=4.1.2

# RAG / AI Assistant
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0
langchain-core>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
